Make turnout and participation floats. Fix tokenDecimals on castedWeight/totalWeight

+++ types.go
@@ -49,7 +49,7 @@ type ElectionInfo struct {
 	CastedVotes             uint64    `json:"voteCount"`
 	CastedWeight            string    `json:"castedWeight,omitempty"`
 	CensusParticipantsCount uint64    `json:"censusParticipantsCount"`
-	Turnout                 int       `json:"turnout"`
+	Turnout                 float32   `json:"turnout"`
 	Username                string    `json:"createdByUsername,omitempty"`
 	Displayname             string    `json:"createdByDisplayname,omitempty"`
 	TotalWeight             string    `json:"totalWeight,omitempty"`
